fixing build for testing
